feat: process booking payment

+++ src/main/java/org/foodie_tour/modules/vnpay/dto/request/PaymentRequest.java
@@ -10,6 +10,6 @@ import lombok.experimental.FieldDefaults;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class PaymentRequest {
-    int bookingId;
-    int amount;
+    long bookingId;
+    long amount;
 }
